pyproject.toml
[project]
name = "litellm-oci"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "litellm[proxy]>=1.79.1",
    "oci>=2.163.0",
    "openai>=1.0.0",
    "pyyaml>=6.0",
]
